SQLite: Revert connection pool PR
Since merging, the CI fail rate increased significantly from ericsink/SQLitePCL.raw#430

Fixes #16202, unresolves #13837
