+++ src/EFCore.Sqlite.Core/Storage/Internal/SqliteDatabaseCreator.cs
@@ -147,7 +147,6 @@ namespace Microsoft.EntityFrameworkCore.Sqlite.Storage.Internal
 
             if (!string.IsNullOrEmpty(path))
             {
-                SqliteConnection.ClearPool(new SqliteConnection(Dependencies.Connection.ConnectionString));
                 File.Delete(path);
             }
         }
