+++ test/Microsoft.Data.Sqlite.Tests/SqliteCommandTest.cs
@@ -926,7 +926,6 @@ namespace Microsoft.Data.Sqlite
             }
             finally
             {
-                SqliteConnection.ClearPool(new SqliteConnection(connectionString));
                 File.Delete("busy.db");
             }
         }
