+++ test/Microsoft.Data.Sqlite.Tests/SqliteConnectionTest.cs
@@ -363,8 +363,6 @@ namespace Microsoft.Data.Sqlite
             }
             finally
             {
-                SqliteConnection.ClearPool(new SqliteConnection("Data Source=encrypted2.db;Password=password"));
-                SqliteConnection.ClearPool(new SqliteConnection("Data Source=encrypted2.db"));
                 File.Delete("encrypted2.db");
             }
         }
